WIP: Add Help>Start Legacy ImageJ 1.x Mode
Since it is very hard (or might be impossible) to achieve full 100%
backwards-compatibility with ImageJ1, we want to have one last way out for
users whose plugin/macro might not work in ImageJ2, namely to show the old
ImageJ1 GUI and lose all the ImageJ2 goodness.

This is what Help>Start Legacy ImageJ 1.x Mode does.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>