WIP: Introduce a legacy mode to the LegacyService
To fall back to original ImageJ1 (e.g. should some plugin or macro not
work in ImageJ2 as it did in ImageJ1), we'd like to be able to drive a
patched ImageJ1 in such a way that it behaves exactly like unpatched
ImageJ1 (except still synchronizing all the data with ImageJ2).

This change introduces a flag in the class LegacyService, and changes the
Javassist patches so that the legacy mode is heeded.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>