Fix image display in the tutorials
Before reworking the way UIs are initialized and picked (to be able to
have multiple UIs in the class path and pick your preference), images
could be displayed without a main window. But now the UIs are asked
which one is visible and if there is none, nothing is shown. That breaks
image display in the LoadAndDisplayDataset example.

This patch falls back to using the default UI if none is currently
visible.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>